% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_tools.R
\name{as.data.frame.list}
\alias{as.data.frame.list}
\title{Convert a List to a Data Frame}
\usage{
\method{as.data.frame}{list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a list to convert to a data frame.}

\item{row.names}{a vector equal to \code{length(x)} corresponding to the row names.
If \code{NULL}, the row names will be set to \code{names(x)}.}

\item{optional}{not used.}

\item{...}{other parameters passed to \code{\link{data.frame}}.}
}
\value{
a data frame.
}
\description{
Convert a List to a Data Frame
}
\examples{
    test1 <- list( c(a='a',b='b',c='c'), c(a='d',b='e',c='f'))
    as.data.frame(test1)

    test2 <- list( c('a','b','c'), c(a='d',b='e',c='f'))
    as.data.frame(test2)

    test3 <- list('Row1'=c(a='a',b='b',c='c'), 'Row2'=c(var1='d',var2='e',var3='f'))
    as.data.frame(test3)

    test4 <- list('Row1'=letters[1:5], 'Row2'=letters[1:7], 'Row3'=letters[8:14])
    as.data.frame(test4)

    test5 <- list(letters[1:10], letters[11:20])
    as.data.frame(test5)

    test6 <- list(list(letters), letters)
    as.data.frame(test6)
}
\references{
\url{http://stackoverflow.com/questions/4227223/r-list-to-data-frame}
}
\author{
Jason Bryer <<jason@bryer.org>>
}
