% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_tools.R
\name{calc_tax}
\alias{calc_tax}
\title{Calculate Tax Burden}
\usage{
calc_tax(amount, tax_table = NA, total = TRUE)
}
\arguments{
\item{amount}{numeric. The amount of taxable income.}

\item{tax_table}{a data frame containing marinal tax rates.}

\item{total}{logical. Return the total amount paid?}
}
\value{
a numeric sum of taxes paid (if total = T) or a vector of taxes paid at the marginal rate
}
\description{
Calculate Tax Burden
}
\examples{
calc_tax(85000)
}
