% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_tools.R
\name{check_for}
\alias{check_for}
\title{Check if Package is Installed}
\usage{
check_for(pkg, quiet = TRUE)
}
\arguments{
\item{pkg}{character. The name of the package to check for.}

\item{quiet}{logical. Should the function suppress any console messages?}
}
\value{
A logical indicating whether the named package was found in the list of installed packages.
}
\description{
Check if Package is Installed
}
\examples{
\dontrun{
    check_for("kittius")
    check_for("dplyr", quiet = FALSE)
}
}
