% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_with.R
\name{find_deltas}
\alias{find_deltas}
\title{Find Differences Between A Data Set and an Updated Version}
\usage{
find_deltas(
  .master,
  .data,
  .key = "VID",
  return_all = FALSE,
  ignore_col_case = TRUE,
  ...
)
}
\arguments{
\item{.master}{a data.frame or tibble containing the master VLD data to use as a baseline.}

\item{.data}{a data.frame or tibble containing the updated data. At least the key column and one additional column must match the column names for the master data.}

\item{.key}{a character. The name of the column on which to match .data and .master. Must be present in both data frames.}

\item{return_all}{logical. Should the output include all rows and columns in the VLD? if FALSE, will only return the columns in both .data and .master with changed values}

\item{ignore_col_case}{logical. Should the column names be compared with or wthout case? if TRUE, will not match columns of differing cases, eg. Tag != tag. Defaults to FALSE, eg. Tag ~= tag.}

\item{...}{additional arguments}
}
\value{
a data frame with base VLD dat
}
\description{
Find Differences Between A Data Set and an Updated Version
}
\examples{
\dontrun{
my_updates \%>\% find_deltas(VLD_master, return_all = FALSE)
}
}
