% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.R
\name{get_fiscal}
\alias{get_fiscal}
\title{Create Fiscal Month Vector}
\usage{
get_fiscal(start = 10, abb = FALSE)
}
\arguments{
\item{start}{numeric. The month in which the fiscal year begins.}

\item{abb}{logical. Return month abbreviations?}
}
\value{
a character vector of month names in fiscal year order
}
\description{
Create Fiscal Month Vector
}
\examples{
federal_year <- get_fiscal(start = 10)
}
