% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not IN}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{a thingy}

\item{y}{another thingy}
}
\value{
a logical
}
\description{
Not IN
}
\examples{
A <- c("Fido", "Leonidas", "Fuzzy", "Tralfaz")
A[A \%notin\% c("Fido")]
}
