% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{makeRoot}
\alias{makeRoot}
\title{Make n-th Root Function
Function factory to create a n-th root function}
\usage{
makeRoot(n)
}
\arguments{
\item{n}{numeric}
}
\value{
a function which complutes the n-th root of an input
}
\description{
Make n-th Root Function
Function factory to create a n-th root function
}
\examples{
square_root <- makeRoot(2)
cube_root <- makeRoot(3)
}
