% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{makeVecNorm}
\alias{makeVecNorm}
\title{Make L-th Norm Function
Function factory to create the L-th norm}
\usage{
makeVecNorm(L)
}
\arguments{
\item{L}{numeric. The l-norm to create}
}
\value{
a function which calculates the L-th norm of a vector
}
\description{
Make L-th Norm Function
Function factory to create the L-th norm
}
\examples{
l2_norm <- makeVecNorm(L = 2)
l3_norm <- makeVecNorm(L = 3)
}
