% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_tools.R
\name{mutate_cond}
\alias{mutate_cond}
\title{Mutate at Condition}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{a dataframe}

\item{condition}{A condition}

\item{...}{all that jazz}

\item{envir}{the environment to scope}
}
\value{
the altered dataframe
}
\description{
Mutate at Condition
}
