% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.R
\name{no_space}
\alias{no_space}
\title{Remove All Spaces From a String}
\usage{
no_space(x, replacement = "")
}
\arguments{
\item{x}{a string}

\item{replacement}{a string to sub for spaces}
}
\value{
a string with spaces removed
}
\description{
Remove All Spaces From a String
}
\examples{
no_space("abcd efgh")
no_space("abcd efgh", replacement = "_")
}
