% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.R
\name{search_code}
\alias{search_code}
\title{Find a Pattern in an .R Code File}
\usage{
search_code(.file, pattern, quiet = TRUE, ...)
}
\arguments{
\item{.file}{character. The file to read in and search.}

\item{pattern}{character. A pattern to pass to grep, in RegExp format.}

\item{quiet}{logical. Should a warning be thrown if the pattern is not in the file?}

\item{...}{pass through additional options}
}
\value{
a data frame containing the line number and line text where the pattern was found in the file.
}
\description{
Reads in a .R code file by line and searches for a given pattern.
}
\examples{
\dontrun{
    search_code("./R/misc_toolr.R", "export")
}
}
