% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_tools.R
\name{smoosh}
\alias{smoosh}
\title{Smoosh Rows in a Data Frame}
\usage{
smoosh(.data, by, to = NULL, ...)
}
\arguments{
\item{.data}{data frame.}

\item{by}{character. The column to use as a key col.}

\item{to}{character vector. The columns to smoosh.}

\item{...}{all that jazz}
}
\value{
a data frame.
}
\description{
Smoosh Rows in a Data Frame
}
