% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax2019.R
\docType{data}
\name{tax2019}
\alias{tax2019}
\title{US Tax Brackets and Merginal Rates for 2019.}
\format{A data frame with 7 rows and 3 variable:
\describe{
  \item{lower}{the lower bound for a tax bracket}
  \item{upper}{the upper bound for a tax bracket}
  \item{rate}{the marginal tax rate for a bracket}
  ...
}}
\source{
\url{https://taxfoundation.org/2019-tax-brackets/}
}
\usage{
tax2019
}
\description{
A dataset containing the US personal tax bracket cutoff limits and
the marginal tax rates for each bracket in 2019.
}
\keyword{datasets}
