% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_with.R
\name{update_with}
\alias{update_with}
\title{Update Data Frame with Changed Fields}
\usage{
update_with(.master, .data, .key = "VID", na_value = NA,
  ignore_col_case = TRUE, replace_names = FALSE, return_list = FALSE,
  ...)
}
\arguments{
\item{.master}{data.frame. The initial data set.}

\item{.data}{data.frame with at least one column name shared with .master}

\item{.key}{character. The name of the shared key/index column between the data frames.}

\item{na_value}{what value should replace NA values in the data?}

\item{ignore_col_case}{logical. Should case be ignored when matching column names?}

\item{replace_names}{logical. Should the lowrcase names unsed in the function be replaced with the original data frame names?}

\item{return_list}{logical. Should a list with both the original data and updated data frames be returned? Defaults to FALSE.}

\item{...}{additional arguments}
}
\value{
either a data.frame of a list of data.frames
}
\description{
Update Data Frame with Changed Fields
}
\examples{
\dontrun{vld \%>\% update_with(new_data1)}
}
